requirements/base.txt
# Base requirements for SGII
# These are the core dependencies needed for the application to run

# Django and core extensions
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
django-rest-framework-nested==0.0.1
django-ratelimit==4.1.0
whitenoise==6.9.0
django-fsm==3.0.0

# Redis support
django-redis==5.4.0

# Database
psycopg2-binary==2.9.10
redis[hiredis]==6.2.0

# Celery for async tasks
celery==5.5.2
amqp==5.3.1
billiard==4.2.1
kombu==5.5.3
vine==5.1.0

# File handling
python-magic==0.4.27
libmagic==1.0
Pillow==10.4.0
reportlab==4.2.5

# AWS S3 support
boto3==1.38.16
botocore==1.38.16
s3transfer==0.12.0

# Utilities
asgiref==3.8.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
jmespath==1.0.1
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13

# Basic setup tools
setuptools==78.1.1
wheel==0.45.1
